Move Channels::get() to channels.rs
